refactor(cli): extract context.ts and term.ts from qmd.ts (batch 1)
Move the two real state clusters out of the monolith, per
docs/specs/qmd-cli-file-split.md:

- context.ts owns store/index lifecycle: getStore/getDb/resyncConfig/
  closeDb/getDbPath/setIndexName/ensureVecTable plus CLI model
  resolution. Adds setStoreDbPathOverride setter for the two call sites
  that wrote the module variable directly.
- term.ts owns terminal primitives: colors, cursor, OSC progress,
  time/byte formatters. SIGINT/SIGTERM cursor-restore handlers move
  with it and keep their import-time registration.

qmd.ts re-exports resolve{Embed,Generate,Rerank}ModelForCli (I4).
Verified: tsc clean, build ok, 32 output snapshots byte-identical,
cli/lazy-llm-import/exit-lifecycle tests green (186 tests total).
Also masks the volatile git commit hash in the version snapshot.
