+++ src/cli/term.ts
@@ -1,0 +1,141 @@
+/**
+ * term.ts - terminal primitives for the qmd CLI: colors, cursor, OSC progress,
+ * and small time/byte formatters.
+ *
+ * Side effects at import time:
+ *   - `useColor` / `isTTY` are evaluated once from process env/stdio.
+ *   - SIGINT/SIGTERM handlers are registered to restore the cursor on exit.
+ *     The foreground HTTP MCP server relies on being the only other code path
+ *     that touches these signals: it calls removeAllListeners("SIGTERM"/
+ *     "SIGINT") before starting, so this module must remain the ONLY src/cli
+ *     module that registers process-level signal handlers at import time.
+ */
+import { homedir } from "../store.js";
+
+// Terminal colors (respects NO_COLOR env)
+const useColor = !process.env.NO_COLOR && process.stdout.isTTY;
+const c = {
+  reset: useColor ? "\x1b[0m" : "",
+  dim: useColor ? "\x1b[2m" : "",
+  bold: useColor ? "\x1b[1m" : "",
+  cyan: useColor ? "\x1b[36m" : "",
+  yellow: useColor ? "\x1b[33m" : "",
+  green: useColor ? "\x1b[32m" : "",
+  magenta: useColor ? "\x1b[35m" : "",
+  blue: useColor ? "\x1b[34m" : "",
+};
+
+// Terminal cursor control
+const cursor = {
+  hide() { process.stderr.write('\x1b[?25l'); },
+  show() { process.stderr.write('\x1b[?25h'); },
+};
+
+// Ensure cursor is restored on exit
+process.on('SIGINT', () => { cursor.show(); process.exit(130); });
+process.on('SIGTERM', () => { cursor.show(); process.exit(143); });
+
+// Terminal progress bar using OSC 9;4 escape sequence (TTY only)
+const isTTY = process.stderr.isTTY;
+const progress = {
+  set(percent: number) {
+    if (isTTY) process.stderr.write(`\x1b]9;4;1;${Math.round(percent)}\x07`);
+  },
+  clear() {
+    if (isTTY) process.stderr.write(`\x1b]9;4;0\x07`);
+  },
+  indeterminate() {
+    if (isTTY) process.stderr.write(`\x1b]9;4;3\x07`);
+  },
+  error() {
+    if (isTTY) process.stderr.write(`\x1b]9;4;2\x07`);
+  },
+};
+
+// Format seconds into human-readable ETA
+function formatETA(seconds: number): string {
+  if (seconds < 60) return `${Math.round(seconds)}s`;
+  if (seconds < 3600) return `${Math.floor(seconds / 60)}m ${Math.round(seconds % 60)}s`;
+  return `${Math.floor(seconds / 3600)}h ${Math.floor((seconds % 3600) / 60)}m`;
+}
+
+function formatTimeAgo(date: Date): string {
+  const seconds = Math.floor((Date.now() - date.getTime()) / 1000);
+  if (seconds < 60) return `${seconds}s ago`;
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes}m ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `${hours}h ago`;
+  const days = Math.floor(hours / 24);
+  return `${days}d ago`;
+}
+
+function formatMs(ms: number): string {
+  if (ms < 1000) return `${ms}ms`;
+  return `${(ms / 1000).toFixed(1)}s`;
+}
+
+function formatBytes(bytes: number): string {
+  if (bytes < 1024) return `${bytes} B`;
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+  if (bytes < 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+  return `${(bytes / (1024 * 1024 * 1024)).toFixed(1)} GB`;
+}
+
+function renderProgressBar(percent: number, width: number = 30): string {
+  const filled = Math.round((percent / 100) * width);
+  const empty = width - filled;
+  const bar = "█".repeat(filled) + "░".repeat(empty);
+  return bar;
+}
+
+// Highlight query terms in text (skip short words < 3 chars)
+function highlightTerms(text: string, query: string): string {
+  if (!useColor) return text;
+  const terms = query.toLowerCase().split(/\s+/).filter(t => t.length >= 3);
+  let result = text;
+  for (const term of terms) {
+    const regex = new RegExp(`(${term.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')})`, 'gi');
+    result = result.replace(regex, `${c.yellow}${c.bold}$1${c.reset}`);
+  }
+  return result;
+}
+
+// Format score with color based on value
+function formatScore(score: number): string {
+  const pct = (score * 100).toFixed(0).padStart(3);
+  if (!useColor) return `${pct}%`;
+  if (score >= 0.7) return `${c.green}${pct}%${c.reset}`;
+  if (score >= 0.4) return `${c.yellow}${pct}%${c.reset}`;
+  return `${c.dim}${pct}%${c.reset}`;
+}
+
+function formatExplainNumber(value: number): string {
+  return value.toFixed(4);
+}
+
+// Shorten directory path for display - relative to $HOME (used for context paths, not documents)
+function shortPath(dirpath: string): string {
+  const home = homedir();
+  if (dirpath.startsWith(home)) {
+    return '~' + dirpath.slice(home.length);
+  }
+  return dirpath;
+}
+
+export {
+  useColor,
+  c,
+  cursor,
+  isTTY,
+  progress,
+  formatETA,
+  formatTimeAgo,
+  formatMs,
+  formatBytes,
+  renderProgressBar,
+  highlightTerms,
+  formatScore,
+  formatExplainNumber,
+  shortPath,
+};
